requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
numpy==2.2.6
openai==2.15.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
portalocker==3.2.0
protobuf==6.33.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.3.0
python-dotenv==1.2.1
python-multipart==0.0.21
qdrant-client==1.16.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
